backend/requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
pydantic>=2.0.0
